+++ earthdaily/earthdatastore/__init__.py
@@ -32,7 +32,9 @@ def post_query_items(items, query):
                                 items_.append(item)
                                 item_already_checked = True
                     else:
-                        operation = operator.__dict__[v_op](item.properties[k], v_val)
+                        operation = operator.__dict__[v_op](
+                            item.properties[k], v_val
+                        )
                     if operation:
                         items_.append(item)
                         item_already_checked = True
@@ -81,7 +83,9 @@ def enhance_assets(
                 if use_http_url:
                     href = item.assets[asset].to_dict().get("href", {})
                     if href:
-                        items[idx].assets[asset].href = _cloud_path_to_http(href)
+                        items[idx].assets[asset].href = _cloud_path_to_http(
+                            href
+                        )
                 if add_default_scale_factor:
                     scale_factor_collection = (
                         _scales_collections.scale_factor_collections.get(
@@ -103,15 +107,15 @@ def enhance_assets(
                                 .extra_fields["raster:bands"][0]
                                 .get("scale")
                             ):
-                                items[idx].assets[asset].extra_fields["raster:bands"][
-                                    0
-                                ]["scale"] = scales_collection["scale"]
-                                items[idx].assets[asset].extra_fields["raster:bands"][
-                                    0
-                                ]["offset"] = scales_collection["offset"]
-                                items[idx].assets[asset].extra_fields["raster:bands"][
-                                    0
-                                ]["nodata"] = scales_collection["nodata"]
+                                items[idx].assets[asset].extra_fields[
+                                    "raster:bands"
+                                ][0]["scale"] = scales_collection["scale"]
+                                items[idx].assets[asset].extra_fields[
+                                    "raster:bands"
+                                ][0]["offset"] = scales_collection["offset"]
+                                items[idx].assets[asset].extra_fields[
+                                    "raster:bands"
+                                ][0]["nodata"] = scales_collection["nodata"]
 
     return items
 
@@ -126,7 +130,9 @@ def _get_client(config=None):
     auth_url = config("EDS_AUTH_URL")
     secret = config("EDS_SECRET")
     client_id = config("EDS_CLIENT_ID")
-    eds_url = config("EDS_API_URL", "https://api.eds.earthdaily.com/archive/v1/stac/v1")
+    eds_url = config(
+        "EDS_API_URL", "https://api.eds.earthdaily.com/archive/v1/stac/v1"
+    )
     if auth_url is None or secret is None or client_id is None:
         raise AttributeError(
             "You need to have env : EDS_AUTH_URL, EDS_SECRET and EDS_CLIENT_ID"
@@ -166,7 +172,10 @@ class StacCollectionExplorer:
 
     @property
     def item_properties(self):
-        return {k: self.item.properties[k] for k in sorted(self.item.properties.keys())}
+        return {
+            k: self.item.properties[k]
+            for k in sorted(self.item.properties.keys())
+        }
 
     def assets(self, asset_name=None):
         if asset_name:
@@ -340,9 +349,9 @@ class Auth:
         """
         if collection:
             if collection not in self._staccollectionexplorer.keys():
-                self._staccollectionexplorer[collection] = StacCollectionExplorer(
-                    self.client, collection
-                )
+                self._staccollectionexplorer[
+                    collection
+                ] = StacCollectionExplorer(self.client, collection)
             return self._staccollectionexplorer.get(collection)
         return sorted(c.id for c in self.client.get_all_collections())
 
@@ -368,10 +377,18 @@ class Auth:
                 )
             if isinstance(collections, list):
                 if len(collections) > 1:
-                    raise ValueError("Mask_with only manage one collection at a time.")
+                    raise ValueError(
+                        "Mask_with only manage one collection at a time."
+                    )
                 collection = collections[0]
             else:
                 collection = collections
+
+        if mask_with == "ag_cloud_mask":
+            # to get only items that have a ag_cloud_mask
+            query = search_kwargs.get("query", {})
+            query.update({"eda:ag_cloud_mask_available": {"eq": True}})
+            search_kwargs["query"] = query
 
         items = self.search(
             collections=collections,
@@ -388,7 +405,9 @@ class Auth:
         )
         if mask_with:
             if mask_with == "native":
-                mask_with = mask._native_mask_def_mapping.get(collection, None)
+                mask_with = mask._native_mask_def_mapping.get(
+                    collection, None
+                )
                 if mask_with is None:
                     raise ValueError(
                         f"Sorry, there's no native mask available for {collection}. Only these collections have native cloudmask : {list(mask._native_mask_mapping.keys())}."
@@ -406,6 +425,7 @@ class Auth:
                 )
                 xr_datacube["time"] = xr_datacube.time.astype("M8[s]")
                 acm_datacube["time"] = acm_datacube.time.astype("M8[s]")
+
                 mask_kwargs.update(acm_datacube=acm_datacube)
             else:
                 mask_assets = mask._native_mask_asset_mapping[collections]
@@ -576,7 +596,9 @@ class Auth:
             for item in items:
                 if not item.properties.get("eda:ag_cloud_mask_available"):
                     continue
-                collection = item.properties["eda:ag_cloud_mask_collection_id"]
+                collection = item.properties[
+                    "eda:ag_cloud_mask_collection_id"
+                ]
                 if products.get(collection, None) is None:
                     products[collection] = []
                 products[collection].append(